Database Design/requirements.txt
Flask>=1.1.2
psycopg2>=2.8.6
python-dotenv>=0.16.0
flask-cors>=3.0.10
pandas>=1.2.4
sqlalchemy>=1.4.8